Move TrieNode to a standalone class and make some fields final.
